+++ apps/examples/src/5-exploded/ExplodedExample.tsx
@@ -1,11 +1,11 @@
-import { Canvas, ContextMenu, TldrawEditor, TldrawUi } from '@tldraw/tldraw'
+import { Canvas, ContextMenu, TldrawEditor, TldrawUi, defaultShapes } from '@tldraw/tldraw'
 import '@tldraw/tldraw/editor.css'
 import '@tldraw/tldraw/ui.css'
 
 export default function ExplodedExample() {
 	return (
 		<div className="tldraw__editor">
-			<TldrawEditor autoFocus persistenceKey="exploded-example">
+			<TldrawEditor autoFocus persistenceKey="exploded-example" shapes={defaultShapes}>
 				<TldrawUi>
 					<ContextMenu>
 						<Canvas />
