+++ packages/editor/src/lib/editor/Editor.ts
@@ -129,9 +129,9 @@ import { getStraightArrowInfo } from './shapes/ArrowShape/ArrowShapeUtil/arrow/s
 import { TLArrowShape } from './shapes/ArrowShape/arrowShapeTypes'
 import { FrameShapeUtil } from './shapes/FrameShape/FrameShapeUtil/FrameShapeUtil'
 import { TLFrameShape } from './shapes/FrameShape/frameShapeTypes'
+import { TextShapeUtil } from './shapes/TextShape/TextShapeUtil/TextShapeUtil'
 import { GroupShapeUtil } from './shapeutils/GroupShapeUtil/GroupShapeUtil'
 import { ShapeUtil, TLResizeMode } from './shapeutils/ShapeUtil'
-import { TextShapeUtil } from './shapeutils/TextShapeUtil/TextShapeUtil'
 import { TLExportColors } from './shapeutils/shared/TLExportColors'
 import { RootState } from './tools/RootState'
 import { StateNode, TLStateNodeConstructor } from './tools/StateNode'
