+++ packages/editor/src/lib/editor/shapes/ArrowShape/ArrowShapeUtil/components/ArrowTextLabel.tsx
@@ -3,7 +3,7 @@ import { TLShapeId } from '@tldraw/tlschema'
 import * as React from 'react'
 import { TEXT_PROPS } from '../../../../../constants'
 import { stopEventPropagation } from '../../../../../utils/dom'
-import { TextHelpers } from '../../../../shapeutils/TextShapeUtil/TextHelpers'
+import { TextHelpers } from '../../../../shapeutils/shared/TextHelpers'
 import { useEditableText } from '../../../../shapeutils/shared/useEditableText'
 import { ARROW_LABEL_FONT_SIZES } from '../../arrowConstants'
 import { TLArrowShape } from '../../arrowShapeTypes'
