+++ packages/editor/src/lib/editor/shapes/TextShape/textShapeTypes.ts
@@ -1,0 +1,16 @@
+import { TLAlignType, TLBaseShape, TLColorType, TLFontType, TLSizeType } from '@tldraw/tlschema'
+
+/** @public */
+export type TLTextShapeProps = {
+	color: TLColorType
+	size: TLSizeType
+	font: TLFontType
+	align: TLAlignType
+	w: number
+	text: string
+	scale: number
+	autoSize: boolean
+}
+
+/** @public */
+export type TLTextShape = TLBaseShape<'text', TLTextShapeProps>
