+++ packages/editor/src/lib/editor/shapeutils/shared/TextHelpers.ts
