+++ packages/tlschema/src/index.ts
@@ -85,7 +85,6 @@ export {
 export { type TLGroupShape } from './shapes/TLGroupShape'
 export { type TLIconShape } from './shapes/TLIconShape'
 export { type TLImageCrop, type TLImageShape, type TLImageShapeProps } from './shapes/TLImageShape'
-export { type TLTextShape, type TLTextShapeProps } from './shapes/TLTextShape'
 export { type TLVideoShape } from './shapes/TLVideoShape'
 export {
 	TL_ALIGN_TYPES,
